+++ components/vf-chatbot-prompt/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.0.0-beta.2
 
 * Updated package.json with correct settings
+* Added : React support [Tracking issue](https://github.com/visual-framework/vf-core/issues/2314)
 
 ### 1.0.0-beta.1