+++ components/vf-chatbot-standalone/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updated : Added more config options
 * Updated : Changes for chatbot modal variant
 * Fixed: Added code to return feedback comments
+* Added : React support [Tracking issue](https://github.com/visual-framework/vf-core/issues/2314)
 
 ### 1.0.0-beta.2