+++ components/vf-chatbot/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updated : Standalone and modal variant initialization moved to `vf-chatbot` component
 * Updated : Added more config options
 * Added : Changes for chatbot modal variant
+* Added : React support [Tracking issue](https://github.com/visual-framework/vf-core/issues/2314)
 
 ### 1.0.0-beta.2